requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu131
torch==2.5.1
torchvision
torchaudio

transformers==4.57.1
datasets
accelerate==1.3.0
peft==0.14.0
trl>=0.8.0

bitsandbytes
scipy
sentencepiece
protobuf
tqdm

wandb==0.22.3
hydra-core==1.3.2
tensorboard

ipykernel
ipywidgets
matplotlib==3.10.8
